moved JsonReader to new package
